C prototipo/Frontend/requirements.txt
Django>=4.2
django-environ>=0.9.0
